Integrate Apr 28 (Base of 0.66) Nightly RN Build
Commits: https://github.com/facebook/react-native/compare/5127c71d6...7e05480cc
