+++ packages/@office-iss/react-native-win32-tester/package.json
@@ -15,7 +15,7 @@
   },
   "peerDependencies": {
     "@office-iss/react-native-win32": "^0.0.0-canary.89",
-    "react-native": "0.0.0-5127c71d6"
+    "react-native": "0.0.0-7e05480cc"
   },
   "devDependencies": {
     "@office-iss/react-native-win32": "^0.0.0-canary.89",
@@ -24,7 +24,7 @@
     "@types/node": "^14.14.22",
     "eslint": "7.12.0",
     "just-scripts": "^1.3.3",
-    "react-native": "0.0.0-5127c71d6",
+    "react-native": "0.0.0-7e05480cc",
     "react-native-platform-override": "^1.4.13",
     "typescript": "^3.8.3"
   }
