+++ packages/@office-iss/react-native-win32/overrides.json
@@ -3,7 +3,7 @@
     ".flowconfig",
     "src/**"
   ],
-  "baseVersion": "0.0.0-5127c71d6",
+  "baseVersion": "0.0.0-7e05480cc",
   "overrides": [
     {
       "type": "derived",
@@ -15,7 +15,7 @@
       "type": "derived",
       "file": "src/index.win32.js",
       "baseFile": "index.js",
-      "baseHash": "377a293707e3a0d18d920af383490957b2e145a5"
+      "baseHash": "f0b44bc2db7a958ed54e45131169b3f6734890f2"
     },
     {
       "type": "patch",
@@ -56,7 +56,7 @@
       "type": "copy",
       "file": "src/Libraries/Components/DatePicker/DatePickerIOS.win32.js",
       "baseFile": "Libraries/Components/DatePicker/DatePickerIOS.android.js",
-      "baseHash": "218895e2a5f3d7614a2cb577da187800857850ea",
+      "baseHash": "a9a4a6c92de922a11f8b81642b1bf1270d0b7b96",
       "issue": 4378
     },
     {
@@ -113,6 +113,7 @@
       "type": "derived",
       "file": "src/Libraries/Components/ProgressViewIOS/ProgressViewIOS.win32.js",
       "baseFile": "Libraries/Components/ProgressViewIOS/ProgressViewIOS.android.js",
+      "baseVersion": "0.0.0-5127c71d6",
       "baseHash": "f0b35fdee6b1c9e7bfe860a9e0aefc00337ea7fd"
     },
     {
@@ -125,7 +126,7 @@
       "type": "copy",
       "file": "src/Libraries/Components/SegmentedControlIOS/SegmentedControlIOS.js",
       "baseFile": "Libraries/Components/SegmentedControlIOS/SegmentedControlIOS.android.js",
-      "baseHash": "37cea8b532ea4e4335120c6f8b2d3d3548e31b18",
+      "baseHash": "6cc05293331f46ca64f1e70d0fa537c5685d3af3",
       "issue": 4378
     },
     {
@@ -152,6 +153,7 @@
       "type": "derived",
       "file": "src/Libraries/Components/TextInput/TextInput.win32.tsx",
       "baseFile": "Libraries/Components/TextInput/TextInput.js",
+      "baseVersion": "0.0.0-5127c71d6",
       "baseHash": "506d84a349fef1ad532b7c02a1836fce7befa616"
     },
     {
@@ -195,13 +197,13 @@
       "type": "patch",
       "file": "src/Libraries/Components/View/ReactNativeViewAttributes.win32.js",
       "baseFile": "Libraries/Components/View/ReactNativeViewAttributes.js",
-      "baseHash": "0825b0257e19cfef2d626f19d219256a01c54b67"
+      "baseHash": "9479c8ccfeeadc90721ae2d3327eeaa76818bfaf"
     },
     {
       "type": "patch",
       "file": "src/Libraries/Components/View/ReactNativeViewViewConfig.win32.js",
       "baseFile": "Libraries/Components/View/ReactNativeViewViewConfig.js",
-      "baseHash": "ba99e0917fe2c518eb11e8b640f759f631621b40"
+      "baseHash": "7d6114901e76fa8f73f84dbb3f486ea2e1c9ce0e"
     },
     {
       "type": "platform",
@@ -232,7 +234,7 @@
       "type": "derived",
       "file": "src/Libraries/Image/Image.win32.js",
       "baseFile": "Libraries/Image/Image.ios.js",
-      "baseHash": "bb70db26cb8d42b51037d43a6ab9eace183a0f71",
+      "baseHash": "c20b0134e74a8e8d69d0b2694b4296fe2bf69d3f",
       "issue": 4320
     },
     {
@@ -243,7 +245,7 @@
       "type": "derived",
       "file": "src/Libraries/Image/NativeImageLoaderWin32.js",
       "baseFile": "Libraries/Image/NativeImageLoaderIOS.js",
-      "baseHash": "fde33e1ab6ab396a91bf469da0aa4bbb4f639383",
+      "baseHash": "3f2be4d90a74ceb6e023dc46fcd0cf3d5a7c47ad",
       "issue": 4320
     },
     {
