+++ packages/@office-iss/react-native-win32/package.json
@@ -66,14 +66,14 @@
     "just-scripts": "^1.3.3",
     "prettier": "1.19.1",
     "react": "17.0.2",
-    "react-native": "0.0.0-5127c71d6",
+    "react-native": "0.0.0-7e05480cc",
     "react-native-platform-override": "^1.4.13",
     "react-shallow-renderer": "16.14.1",
     "typescript": "^3.8.3"
   },
   "peerDependencies": {
     "react": "17.0.2",
-    "react-native": "0.0.0-5127c71d6"
+    "react-native": "0.0.0-7e05480cc"
   },
   "beachball": {
     "defaultNpmTag": "canary",
