+++ packages/@office-iss/react-native-win32/src/Libraries/Components/DatePicker/DatePickerIOS.win32.js
@@ -8,11 +8,10 @@
  */
 
 'use strict';
-
-const React = require('react');
-const StyleSheet = require('../../StyleSheet/StyleSheet');
-const Text = require('../../Text/Text');
-const View = require('../View/View');
+import * as React from 'react';
+import StyleSheet from '../../StyleSheet/StyleSheet';
+import Text from '../../Text/Text';
+import View from '../View/View';
 
 class DummyDatePickerIOS extends React.Component {
   render() {
