+++ packages/@office-iss/react-native-win32/src/Libraries/Image/NativeImageLoaderWin32.js
@@ -8,6 +8,7 @@
  * @format
  */
 
+import type {RootTag} from '../TurboModule/RCTExport';
 import type {TurboModule} from '../TurboModule/RCTExport';
 import * as TurboModuleRegistry from '../TurboModule/TurboModuleRegistry';
 
@@ -34,7 +35,7 @@ export interface Spec extends TurboModule {
   +prefetchImageWithMetadata?: (
     uri: string,
     queryRootName: string,
-    rootTag: number,
+    rootTag: RootTag,
   ) => Promise<boolean>;
   +queryCache: (uris: Array<string>) => Promise<Object>;
   // Win32]
