+++ packages/@react-native-windows/tester/package.json
@@ -14,7 +14,7 @@
     "@react-native/tester": "0.0.1"
   },
   "peerDependencies": {
-    "react-native": "0.0.0-5127c71d6",
+    "react-native": "0.0.0-7e05480cc",
     "react-native-windows": "^0.0.0-canary.314"
   },
   "devDependencies": {
@@ -23,7 +23,7 @@
     "@types/node": "^14.14.22",
     "eslint": "7.12.0",
     "just-scripts": "^1.3.3",
-    "react-native": "0.0.0-5127c71d6",
+    "react-native": "0.0.0-7e05480cc",
     "react-native-platform-override": "^1.4.13",
     "react-native-windows": "^0.0.0-canary.314",
     "typescript": "^3.8.3"
