+++ packages/@react-native/repo-config/package.json
@@ -34,7 +34,7 @@
     "eslint-plugin-react-native": "3.10.0",
     "eslint-plugin-relay": "1.8.1",
     "flow-bin": "^0.149.0",
-    "jest": "^26.5.2",
+    "jest": "^26.6.3",
     "jest-junit": "^10.0.0",
     "jscodeshift": "^0.11.0",
     "mkdirp": "^0.5.1",
