+++ packages/@rnw-bots/coordinator/package.json
@@ -21,7 +21,7 @@
     "@types/node": "^14.14.22",
     "babel-jest": "^26.3.0",
     "eslint": "7.12.0",
-    "jest": "^26.5.2",
+    "jest": "^26.6.3",
     "just-scripts": "^1.3.3",
     "prettier": "1.19.1",
     "typescript": "^3.8.3"
