+++ packages/@rnw-scripts/doxysaurus/package.json
@@ -37,7 +37,7 @@
     "@types/yargs": "^16.0.0",
     "babel-jest": "^26.3.0",
     "eslint": "7.12.0",
-    "jest": "^26.5.2",
+    "jest": "^26.6.3",
     "just-scripts": "^1.3.3",
     "prettier": "1.19.1",
     "typescript": "^3.8.3"
