+++ packages/@rnw-scripts/jest-debug-config/package.json
@@ -12,7 +12,7 @@
     "@babel/preset-typescript": "^7.8.3",
     "babel-jest": "^26.3.0",
     "eslint": "7.12.0",
-    "jest": "^26.5.2",
+    "jest": "^26.6.3",
     "prettier": "1.19.1"
   },
   "peerDependencies": {
