+++ packages/e2e-test-app/package.json
@@ -14,7 +14,7 @@
     "@react-native-windows/tester": "0.0.1",
     "node-rnw-rpc": "^1.0.6-12",
     "react": "17.0.2",
-    "react-native": "0.0.0-5127c71d6",
+    "react-native": "0.0.0-7e05480cc",
     "react-native-windows": "^0.0.0-canary.314"
   },
   "devDependencies": {
@@ -31,7 +31,7 @@
     "@types/react-native": "^0.64.4",
     "babel-jest": "^26.3.0",
     "eslint": "7.12.0",
-    "jest": "^26.5.2",
+    "jest": "^26.6.3",
     "jest-environment-winappdriver": "^1.1.6-15",
     "just-scripts": "^1.3.3",
     "metro-react-native-babel-preset": "^0.56.0",
