+++ packages/integration-test-app/package.json
@@ -13,7 +13,7 @@
     "chai": "^4.2.0",
     "node-rnw-rpc": "^1.0.6-12",
     "react": "17.0.2",
-    "react-native": "0.0.0-5127c71d6",
+    "react-native": "0.0.0-7e05480cc",
     "react-native-windows": "^0.0.0-canary.314"
   },
   "devDependencies": {
@@ -33,7 +33,7 @@
     "@types/ws": "^7.4.0",
     "babel-jest": "^26.3.0",
     "eslint": "7.12.0",
-    "jest": "^26.5.2",
+    "jest": "^26.6.3",
     "just-scripts": "^1.3.3",
     "ora": "^3.4.0",
     "prettier": "1.19.1",
