+++ packages/node-rnw-rpc/package.json
@@ -27,7 +27,7 @@
     "just-scripts": "^1.3.2",
     "prettier": "1.19.1",
     "react": "17.0.2",
-    "react-native": "0.0.0-5127c71d6",
+    "react-native": "0.0.0-7e05480cc",
     "react-native-windows": "^0.0.0-canary.314",
     "typescript": "^3.8.3"
   },
