+++ packages/react-native-platform-override/package.json
@@ -64,7 +64,7 @@
     "babel-jest": "^26.3.0",
     "diff-match-patch": "^1.0.4",
     "eslint": "7.12.0",
-    "jest": "^26.5.2",
+    "jest": "^26.6.3",
     "just-scripts": "^1.3.3",
     "minimatch": "^3.0.4",
     "prettier": "1.19.1",
