+++ packages/sample-apps/package.json
@@ -12,7 +12,7 @@
   },
   "dependencies": {
     "react": "17.0.2",
-    "react-native": "0.0.0-5127c71d6",
+    "react-native": "0.0.0-7e05480cc",
     "react-native-windows": "0.0.0-canary.314"
   },
   "devDependencies": {
